Move inspector.js into components folder
