+++ lib/components/inspector/component.js
