+++ lib/components/inspector/index.js
@@ -3,11 +3,11 @@
 import React from 'react';
 import * as transformime from 'transformime';
 
-import log from './utils/log';
-import store from './store';
-import reactFactory from './utils/react';
-import { getCodeToInspect } from './code-manager';
-import InspectorComponent from './components/inspector';
+import log from './../../utils/log';
+import store from './../../store';
+import reactFactory from './../../utils/react';
+import { getCodeToInspect } from './../../code-manager';
+import InspectorComponent from './component';
 
 const transform = transformime.createTransform();
