+++ lib/main.js
@@ -19,7 +19,7 @@ import disposer from './store/disposer';
 import Config from './config';
 import kernelManager from './kernel-manager';
 import ZMQKernel from './zmq-kernel';
-import Inspector from './inspector';
+import Inspector from './components/inspector';
 import AutocompleteProvider from './autocomplete-provider';
 import HydrogenProvider from './plugin-api/hydrogen-provider';
 import reactFactory from './utils/react';
